frontend/requirements.txt
streamlit==1.28.1
requests==2.31.0
Pillow==11.3.0
pandas==2.1.3
